app/requirements.txt
Flask==3.0.3
google-cloud-pubsub==2.21.1
gunicorn==21.2.0
